test code 수정, 메인 코드 수정

+++ users/services.py
@@ -78,19 +78,19 @@ class UserService:
         try:
             if password == "" or password is None:
                 raise ValidationError("비밀번호 필드에 공백이 입력되었습니다.")
-            if check_password(password,user.password):
+
+            if check_password(password, user.password):
                 with transaction.atomic():
                     s3 = client("s3")
-                    s3.delete_object(
-                        Bucket=os.getenv("AWS_STORAGE_BUCKET_NAME"),
-                        Key=user.profile_image.name,
-                    )
-
+                    if user.profile_image and user.profile_image.name:
+                        s3.delete_object(
+                            Bucket=os.getenv("AWS_STORAGE_BUCKET_NAME"),
+                            Key=user.profile_image.name,
+                        )
                     user.delete()
                     return True
-
         except Exception as e:
-            raise e
+            raise Exception(str(e))
 
     @staticmethod
     def update_user_role(user: User, role: str) -> None:
